requirements.txt
Django>=3.8
psycopg2-binary>=2.8
django-use-email-as-username
django-allauth
django-crispy-forms
pillow
django-ses